Extract the error raised by `restoreT` into a separate definition
This prevents the lengthy string containing the error message from being
included in the unfolding of `restoreT`, since the use of INLINABLE
(used here to ensure cross-module specialization) prevents it from being
floated out. The effect of this is almost certainly completely
inconsequential, but it’s technically the Right Thing to ensure the
unfolding of a binding annotated with INLINE or INLINABLE is as small as
possible, and it certainly can’t hurt.